factor out validations
